pyproject.toml
[project]
name = "ticketingtool"
version = "0.1.0"
description = "A ticketing tool for personal use built on FastAPI and hopefully with a Good UI with JS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.122.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=0.2.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.8",
]
